+++ admin/release.py
@@ -21,6 +21,14 @@ def get_homebrew_formula(version: str) -> str:
     requirements_file = Path(__file__).parent.parent / 'requirements.txt'
     lines = requirements_file.read_text().strip().split('\n')
     requirements = [line for line in lines if not line.startswith('#')]
+    # Keyring is not a direct dependency but without it some users get:
+    #
+    # Cannot load 'keyring' on your system (either not installed, or not
+    # configured correctly): No module named 'keyring'
+    requirements.append('keyring')
+    # Similarly, without the following, some users get:
+    # The 'secretstorage' distribution was not found and is required by keyring
+    requirements.append('secretstorage')
     first = requirements[0]
 
     args = ['poet', first]
@@ -40,6 +48,7 @@ def get_homebrew_formula(version: str) -> str:
           head "https://github.com/dcos/dcos-e2e.git"
           homepage "http://dcos-e2e.readthedocs.io/en/latest/cli.html"
           depends_on "python3"
+          depends_on "pkg-config"
 
         {resource_stanzas}