+++ src/cli/_common.py
@@ -21,6 +21,7 @@ from dcos_e2e.node import Node, Transport
 LINUX_DISTRIBUTIONS = {
     'centos-7': Distribution.CENTOS_7,
     'coreos': Distribution.COREOS,
+    'ubuntu-16.04': Distribution.UBUNTU_16_04,
 }
 
 DOCKER_VERSIONS = {
@@ -33,11 +34,6 @@ DOCKER_STORAGE_DRIVERS = {
     'aufs': DockerStorageDriver.AUFS,
     'overlay': DockerStorageDriver.OVERLAY,
     'overlay2': DockerStorageDriver.OVERLAY_2,
-}
-
-TRANSPORTS = {
-    'ssh': Transport.SSH,
-    'docker-exec': Transport.DOCKER_EXEC,
 }
 
 CLUSTER_ID_LABEL_KEY = 'dcos_e2e.cluster_id'