+++ src/cli/commands/web.py
@@ -1,0 +1,29 @@
+"""
+Tools for opening a cluster's web UI.
+"""
+
+import click
+
+from cli._common import ClusterContainers
+from cli._options import existing_cluster_id_option
+from dcos_e2e.node import Transport
+
+
+@click.command('web')
+@existing_cluster_id_option
+def web(cluster_id: str) -> None:
+    """
+    Open the browser at the web UI.
+
+    Note that the web UI may not be available at first.
+    Consider using ``dcos-docker wait`` before running this command.
+    """
+    cluster_containers = ClusterContainers(
+        cluster_id=cluster_id,
+        # The transport is not used so does not matter.
+        transport=Transport.DOCKER_EXEC,
+    )
+    cluster = cluster_containers.cluster
+    master = next(iter(cluster.masters))
+    web_ui = 'http://' + str(master.public_ip_address)
+    click.launch(web_ui)