+++ src/dcos_e2e/backends/_docker/_containers.py
@@ -82,8 +82,6 @@ def start_dcos_container(
     container_number: int,
     mounts: List[docker.types.Mount],
     tmpfs: Dict[str, str],
-    dcos_num_masters: int,
-    dcos_num_agents: int,
     docker_image: str,
     labels: Dict[str, str],
     public_key_path: Path,
@@ -105,10 +103,6 @@ def start_dcos_container(
             http://docker-py.readthedocs.io/en/latest/containers.html.
         tmpfs: See `tmpfs` on
             http://docker-py.readthedocs.io/en/latest/containers.html.
-        dcos_num_masters: The number of master nodes expected to be in the
-            cluster once it has been created.
-        dcos_num_agents: The number of agent nodes (agent and public
-            agents) expected to be in the cluster once it has been created.
         docker_image: The name of the Docker image to use.
         labels: Docker labels to add to the cluster node containers. Akin to
             the dictionary option in
@@ -125,11 +119,7 @@ def start_dcos_container(
     else:
         extra_host_ip_address = '127.0.0.1'
     hostname = container_base_name + str(container_number)
-    environment = {
-        'container': hostname,
-        'DCOS_NUM_MASTERS': dcos_num_masters,
-        'DCOS_NUM_AGENTS': dcos_num_agents,
-    }
+    environment = {'container': hostname}
     extra_hosts = {registry_host: extra_host_ip_address}
 
     client = docker.from_env(version='auto')
@@ -181,11 +171,9 @@ def start_dcos_container(
         ['mkdir', '--parents', '/root/.ssh'],
         '/bin/bash -c "{cmd}"'.format(cmd=' '.join(echo_key)),
         ['rm', '-f', '/run/nologin', '||', 'true'],
-        ['systemctl', 'enable', 'sshd'],
         ['systemctl', 'start', 'sshd'],
         # Work around https://jira.mesosphere.com/browse/DCOS_OSS-1361.
         ['systemd-tmpfiles', '--create', '--prefix', '/run/log/journal'],
     ]:
-        container.exec_run(cmd=cmd)
         exit_code, output = container.exec_run(cmd=cmd)
         assert exit_code == 0, ' '.join(cmd) + ': ' + output.decode()