+++ tests/test_dcos_e2e/test_cluster.py
@@ -249,7 +249,7 @@ class TestClusterFromNodes:
     def test_cluster_from_nodes(self, cluster_backend: ClusterBackend) -> None:
         """
         It is possible to create a cluster from existing nodes, but not destroy
-        it.
+        it, or any nodes in it.
         """
         cluster = Cluster(
             cluster_backend=cluster_backend,
@@ -284,6 +284,9 @@ class TestClusterFromNodes:
         with pytest.raises(NotImplementedError):
             duplicate_cluster.destroy()
 
+        with pytest.raises(NotImplementedError):
+            duplicate_cluster.destroy_node(node=duplicate_master)
+
         cluster.destroy()
 
     def test_install_dcos(
@@ -293,32 +296,46 @@ class TestClusterFromNodes:
         cluster_backend: ClusterBackend,
     ) -> None:
         """
-        If a user attempts to install DC/OS on is called on a `Cluster` created
-        from existing nodes, a `NotImplementedError` is raised.
+        DC/OS can be installed on an existing cluster from a URL and not
         """
         with Cluster(
             cluster_backend=cluster_backend,
             masters=1,
             agents=0,
             public_agents=0,
-        ) as cluster:
+        ) as original_cluster:
             cluster = Cluster.from_nodes(
-                masters=cluster.masters,
-                agents=cluster.agents,
-                public_agents=cluster.public_agents,
+                masters=original_cluster.masters,
+                agents=original_cluster.agents,
+                public_agents=original_cluster.public_agents,
             )
-
-            dcos_config = cluster.base_config
-            assert dcos_config == {}
-
-            with pytest.raises(NotImplementedError):
-                cluster.install_dcos_from_url(
-                    build_artifact=oss_artifact_url,
-                    dcos_config=dcos_config,
-                )
 
             with pytest.raises(NotImplementedError):
                 cluster.install_dcos_from_path(
                     build_artifact=oss_artifact,
-                    dcos_config=dcos_config,
+                    dcos_config=cluster.base_config,
                 )
+
+            cluster.install_dcos_from_url(
+                build_artifact=oss_artifact_url,
+                dcos_config=original_cluster.base_config,
+            )
+
+            cluster.wait_for_dcos_oss()
+
+
+class TestDestroyNode:
+    """
+    Tests for destroying nodes.
+    """
+
+    def test_destroy_node(self, cluster_backend: ClusterBackend) -> None:
+        """
+        It is possible to destroy a node in the cluster.
+        """
+        with Cluster(cluster_backend=cluster_backend) as cluster:
+            (agent, ) = cluster.agents
+            cluster.destroy_node(node=agent)
+            assert not cluster.agents
+            with pytest.raises(CalledProcessError):
+                agent.run(args=['echo', '1'])